Extract product related variables from OfferItem into own class
